+++ API/Presentation/API/Controllers/CommentsController.cs
@@ -26,8 +26,8 @@ namespace API.Controllers
             _commentWriteRepository = commentWriteRepository;
             _commentReadRepository = commentReadRepository;
 
-            _userReadRepository = userReadRepository;
             _productReadRepository = productReadRepository;
+            _userReadRepository = userReadRepository;
         }
 
         [HttpGet]
@@ -39,50 +39,82 @@ namespace API.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id)
         {
-            return Ok(await _commentReadRepository.GetByIdAsync(id, false));
+            Comment comment = await _commentReadRepository.GetByIdAsync(id, false);
+            if (comment == null)
+                return NotFound("Comment does not exist.");
+
+            return Ok(comment);
+        }
+
+        [HttpGet("ByProduct/{id}")]
+        public async Task<IActionResult> ByProduct(int id)
+        {
+            if (await _productReadRepository.GetByIdAsync(id, false) == null)
+                return NotFound("Product does not exist.");
+
+            return Ok(_commentReadRepository.GetWhere(comment => comment.ProductId == id, false));
+        }
+
+        [HttpGet("ByUser/{id}")]
+        public async Task<IActionResult> ByUser(int id)
+        {
+            if (await _userReadRepository.GetByIdAsync(id, false) == null)
+                return NotFound("User does not exist.");
+
+            return Ok(_commentReadRepository.GetWhere(comment => comment.UserId == id, false));
         }
 
         [HttpPost]
         public async Task<IActionResult> Post(VM_Create_Comment modelComment)
         {
             Product product = await _productReadRepository.GetByIdAsync(modelComment.ProductId);
+            if (product == null)
+                return NotFound("Product does not exist.");
+
             User user = await _userReadRepository.GetByIdAsync(modelComment.UserId);
+            if (user == null)
+                return NotFound("User does not exist.");
 
             await _commentWriteRepository.AddAsync(new()
             {
                 Title = modelComment.Title,
                 Text = modelComment.Text,
                 Rating = modelComment.Rating,
-                CommentedTo = product,
-                WrittenBy = user
+                Product = product,
+                User = user
             });
 
             await _commentWriteRepository.SaveAsync();
-            return Ok();
+            return Ok("Comment created.");
         }
 
         [HttpPut]
         public async Task<IActionResult> Put(VM_Update_Comment modelComment)
         {
-            Comment category = await _commentReadRepository.GetByIdAsync(modelComment.CommentId);
+            Comment comment = await _commentReadRepository.GetByIdAsync(modelComment.CommentId);
+            if (comment == null)
+                return NotFound("Comment does not exist.");
 
             if (modelComment.Title != null)
-                category.Title = modelComment.Title;
+                comment.Title = modelComment.Title;
             if (modelComment.Text != null)
-                category.Text = modelComment.Text;
+                comment.Text = modelComment.Text;
             if (modelComment.Rating != null)
-                category.Rating = (int)modelComment.Rating;
+                comment.Rating = (int)modelComment.Rating;
 
             await _commentWriteRepository.SaveAsync();
-            return Ok();
+            return Ok("Comment updated.");
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (await _commentReadRepository.GetByIdAsync(id, false) == null)
+                return NotFound("Comment does not exist.");
+
             await _commentWriteRepository.RemoveAsync(id);
             await _commentWriteRepository.SaveAsync();
-            return Ok();
+            return Ok("Comment deleted.");
         }
     }
 }
