+++ API/Presentation/API/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
-﻿using Application.ViewModels;
+﻿using Application.Validators;
+using Application.ViewModels;
 using Application.Repositories;
 using Domain.Entities;
 using Microsoft.AspNetCore.Http;
@@ -15,19 +16,23 @@ namespace API.Controllers
 
         readonly private ILikeReadRepository _likeReadRepository;
         readonly private ILikeWriteRepository _likeWriteRepository;
+        readonly private ISellerReadRepository _sellerReadRepository;
 
         public UsersController(
             IUserWriteRepository userWriteRepository,
             IUserReadRepository userReadRepository,
 
             ILikeReadRepository likeReadRepository,
-            ILikeWriteRepository likeWriteRepository)
+            ILikeWriteRepository likeWriteRepository,
+
+            ISellerReadRepository sellerReadRepository)
         {
             _userWriteRepository = userWriteRepository;
             _userReadRepository = userReadRepository;
 
             _likeReadRepository = likeReadRepository;
             _likeWriteRepository = likeWriteRepository;
+            _sellerReadRepository = sellerReadRepository;
         }
 
         [HttpGet]
@@ -39,56 +44,101 @@ namespace API.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id)
         {
-            return Ok(await _userReadRepository.GetByIdAsync(id, false));
+            User user = await _userReadRepository.GetByIdAsync(id, false);
+            if (user == null)
+                return NotFound("User does not exist.");
+
+            return Ok(user);
         }
 
         [HttpPost]
         public async Task<IActionResult> Post(VM_Create_User modelUser)
         {
+            // Checking if Username and Email are Unique for both User and Seller
+            if (await _userReadRepository.GetSingleAsync(user => user.Username == modelUser.Username) != null)
+                return BadRequest("Username already exists.");
+            if (await _sellerReadRepository.GetSingleAsync(seller => seller.Username == modelUser.Username) != null)
+                return BadRequest("Username already exists.");
+            if (await _userReadRepository.GetSingleAsync(user => user.EMail == modelUser.EMail) != null)
+                return BadRequest("EMail already exists.");
+            if (await _sellerReadRepository.GetSingleAsync(seller => seller.EMail == modelUser.EMail) != null)
+                return BadRequest("EMail already exists.");
+
+            // EMail Validation with Regex
+            if (!EMailValidation.CheckEMail(modelUser.EMail))
+                return BadRequest("Invalid EMail.");
+
+            // Admin Validation with Custom Admin Password
+            bool admin = false;
+            if (AdminValidation.CheckAdmin(modelUser.Admin))
+                admin = true;
+
             await _userWriteRepository.AddAsync(new()
             {
                 Name = modelUser.Name,
                 Username = modelUser.Username,
                 EMail = modelUser.EMail,
                 Password = modelUser.Password,
-                Admin = modelUser.Admin
+                Admin = admin
             });
 
             await _userWriteRepository.SaveAsync();
-            return Ok();
+            return Ok("User created.");
         }
 
         [HttpPut]
         public async Task<IActionResult> Put(VM_Update_User modelUser)
         {
             User user = await _userReadRepository.GetByIdAsync(modelUser.UserId);
+            if (user == null)
+                return NotFound("User does not exist.");
 
             if (modelUser.Name != null)
                 user.Name = modelUser.Name;
             if (modelUser.Username != null)
+            {
+                if (await _userReadRepository.GetSingleAsync(user => user.Username == modelUser.Username) != null)
+                    return BadRequest("Username already exists.");
+
                 user.Username = modelUser.Username;
+            }
             if (modelUser.EMail != null)
+            {
+                if (await _userReadRepository.GetSingleAsync(user => user.EMail == modelUser.EMail) != null)
+                    return BadRequest("EMail already exists.");
+                if (!EMailValidation.CheckEMail(modelUser.EMail))
+                    return BadRequest("Invalid EMail.");
+
                 user.EMail = modelUser.EMail;
+            }
             if (modelUser.Password != null)
                 user.Password = modelUser.Password;
             if (modelUser.Admin != null)
-                user.Admin = (bool)modelUser.Admin;
+            {
+                if (AdminValidation.CheckAdmin(modelUser.Admin))
+                    user.Admin = true;
+
+                else { user.Admin = false; }
+            }
 
             await _userWriteRepository.SaveAsync();
-            return Ok();
+            return Ok("User updated.");
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
-            // Deleting User Likes Manually Because MsSQL Cascade problems.
-            User user = await _userReadRepository.GetByIdAsync(id);
-            var userComments = _likeReadRepository.GetWhere(comment => comment.LikedBy == user).ToList();
-            _likeWriteRepository.RemoveRange(userComments);
+            if (await _userReadRepository.GetByIdAsync(id) == null)
+                return NotFound("User does not exist.");
+
+            // Manually Deleting User Likes First, Because of the MsSQL Cascade problems.
+            var userLikes = _likeReadRepository.GetWhere(likes => likes.UserId == id).ToList();
+            if (userLikes != null)
+                _likeWriteRepository.RemoveRange(userLikes);
 
             await _userWriteRepository.RemoveAsync(id);
             await _userWriteRepository.SaveAsync();
-            return Ok();
+            return Ok("User deleted.");
         }
     }
 }
